Minor changes
a few simple grammatical errors. Figured out how to print the quotes in the most basic way, but still need to have them hold to the CSS. Styling looks terrible. Considering an if/else if/else statement to help rotate in quote, year, citation, etc if the conditions are met in the global array.